requirements.txt
setuptools>=65.0.0
wheel>=0.38.0
beautifulsoup4==4.9.3
Flask==2.3.3
googlesearch_python==1.0.1
numpy==1.24.3
pandas==1.5.3
python_dateutil==2.8.2
requests==2.31.0
scikit_learn==1.3.0
whois==0.9.13
gunicorn==21.2.0
urllib3==1.26.18
